qtenv: Fix `zoomfactor` preference keys by formatting numbers properly
The implicit conversion from char (implicitly narrowed from int)
to QChar was removed in Qt 6. And even before, this was wrong,
as it added the ASCII NUL to the string instead of '0'.
The plain (non-module) canvas inspector/viewer is just so rarely
used anyway, that probably nobody noticed this. This is also
demonstrated by the fact that the more ergonomic `getPref`/`setPref`
methods, which handle this correctly, are not used here.