ide: Added a deadlock detection mechanism and an example that creates a deadlock.
There are two kinds of deadlocks in the IDE:
1. Between threads
2. UI lockup (typically syncExec)

The first kind of deadlock is created between two jobs using two locks, one for the workspace
and another one for the e.g. NED resources. The detector checks for deadlocks periodically
and the resolver interrupts the locking mechanism. Only locks acquired by lockInterruptibly()
can be interrupted this way.

Eclipse itself does not detect deadlocks automatically.

The second kind of deadlock goes e.g. like this: A job calls Display.syncExec()
(which submits a lambda into the UI event queue, and waits until it is
processed). But for some reason, the submitted lambda (or a non-yet
execured lambda submitted by an earlier asyncExec call) needs to lock a
resource that is currently held by the job --> UI locks up.

"Normal" deadlocks are detected by threadMxBean.findDeadlockedThreads().
UI deadlocks are detected by checking for the UI thread being blocked
for a long time (e.g. 10s).

Resolution is to interrupt. But this is only possible if the lock was acquired
with lockInterruptibly(). So we need to change the locking mechanism
whereever possible.

DISABLING: Deadlock detection can be disabled by defining the following env var
to any value: OMNETPP_DISABLE_DEADLOCK_DETECTION