prod: mingwenv.cmd renamed to opp_shell.cmd (mingwenv.cmd also kept for compatibility)
The script launches an msys environment and this is no longer based on the
mingw but rather on the clang64 toolchain. It is also possible to use other
environments (like clangarm64 etc) by changing the ENVIRONMENT variable
before the first installation. In this case also set USE_MSYS_SEED to "yes"
so only a minimal Msys2 environment will be installed on the first start
and then download the necessary dependencies and toolchain with the
install.sh script.