qtenv: Add ElidingLabel for the status line labels
With larger (presentation-compatible) UI font sizes,
these labels often put a minimum width constraint on
the main window, sometimes even forcing it to be
resized while a simulation was runnning, due to the
changing content in them. In extreme cases, this
could even force the main window to always be wider
than the screen, breaking the "maximize" functionality.
Understandably, this was deemed undesirable.
Now the labels will only have a minimum width to
contain the ellipsis ("..."). If there isn't enough
space to display the full text, it will be elided.