qtenv: Rename actions: "Debug Now" -> "Break into Debugger", "Debug Simulation" -> "Debug Next Step"
The "Debug Now" action is useful less often, so a "less
inviting" name was chosen for it. And "Debug Simulation"
was nearly reverted to "Debug Next Event", which it was
called previously, but now that it also breaks when
resuming from a pausePoint(), this is more fitting.
See: 42c21134f4e4e5eb158a42e452cf7536cd74b473

The docs didn't even reflect the previous rename of
"Debug Next Event" -> "Debug Simulation".