build: rewritten dependency check in omnetpp.scave - setuptools is soon to be deprecated
The setuptools module will soon be deprecated. The version check
was reimplemented to use the modern replacement (metadata and packaging)

For now the test works with both the new and the legacy module
first trying to use the new and then the old. If neither of them
is present, it assumes that the dependencies are correctly installed
by the user. (so the direct test for setuptools was removed from
'configure'). The requirements file now installs packaging
instead of setuptools (but we may be using setuptools on Windows)