build: bundle a python3-config script directly instead of using the system package
The python3-config system package is NOT considered as part of the python virtual environment.
This led to numerous issues where the virtual environment was using different python version
than the python3-config system script causing build errors. We keep a python version of
the python3-config script in the 'python' directory and make sure that we use that
for getting any config info. This guaranties that the returned info will be consistent
with the python version used.