qtenv: Rename columns in the log inspector: "Event#" -> "Sender Event#", "Time" -> "Sending Time"
They were somewhat confusing, as they suggested
that the message shown on that line was _processed_
(arrived/executed) as the event with the shown
event number and at the shown simulation time.
And in the "log" view, the same values indeed
appeared as such later (when/if it arrived).

Logically, at least the first one (what will be
the event number of the message that was just
sent, when it arrives) _can't even be known_ at
the time, as other events may come and go between
sending and arrival of any message.