scave: Reduce margins around data on native line plots
In internal testing and usage, the horizontal margin
was deemed to be a waste of screen real estate,
especially considering that they are proportional,
and chart viewports are usually much wider than they
are tall (on common widescreen displays, and without
the chart script editor being open).

The vertical margin was reduced only slightly,
because it was found to be useful more often to
keep the data points at the extrema further away
from the border of the chart area in the value
dimension. For example, when a vector is
alternating between 0 and 1 (or other constant).
If there was no margin, only the transitions
would be visible as vertical lines, but not
really where the value is 0 or not.