ide/ned: fix #1397 "NED editor adds extra display string tag args when changing icon"
Editing with the Property Dialog added ",,0" to the i tag.

After experimenting with several possible solutions we finally replaced spinner editors in the PropertiesDialog with TextFields

Spinners cannot represent properly the data model in the display string because
they cannot represent a "missing" value.
e.g. deleteing the content of the spinner and then tabbing over it will replace "" with 0