ide: Don't link libgcc and libstdc++ statically into nativelibs anymore
This caused sporadic red "<error>" strings to appear in the data tables
of the "Browse Data" tab in the Analysis Tool of the IDE, if a WebKit
view was shown first, dynamically loading another libstdc++ instance
of a different version (and internal ABI) from the system.
In this case, the "C" locale broke due to conflicting std::locale::facet
linker symbols, and as a result, `std::stringstream::operator<<(double)`
broke, making QuantityFormatter::format also fail (throw an exception).

Despite of this, we still use static linking on Windows to reduce dependencies.
It is unlikely that this causes issues on Windows because shared libs used by
the IDE, especially the webbrowser does not use the libstdc++.dll from mingw.