nedxml: fix #1449 "MSG enum class doesn't compile"
Issue: As usage places, generated class descriptor uses enum2string()
takes an int, but an enum class, unlike plain enum, is NOT implicitly
convertible to int. Solution: use static_cast in the generated code.

Interestingly, there was a test for enum @class (test/core/msg_enum_4.test),
but it did not catch the error because it inly contained enum declarations,
but no usages.

KNOWN FURTHER ISSUE: enum2string(), string2enum(), cEnum, etc represent
the value as int. However, enum class allows specifying the underlying type,
and if it is int64_t or something that does not fit into an int,
enum2string() and string2enum() will lose significant bits and won't
work correctly. Resolving this would not be worth the trouble,
because enums of that size are not practical (i.e. nobody will do that),
and even then, the consequence of errors would be mild.