ide/scave: show Console when chart script editor is opened
Rationale: if user wants to look at (or edit) the chart script, it is almost
certain that they will be interested in seeing the console output:
any warnings, result of print() statements, etc.

At the same time, this automatic action is unlikely to be intrusive:
user does not click the "Open chart script" button very often (as it stays
open afterwards), so it does not interfere with the user if they want to
see the contents of another View.