qtenv: Make runs in the Run combo in the Run Selection Dialog filterable by typing
The "Config name" combo was NOT made filterable due to usability reasons.
(Selection from the combo is made harder in an editable combo, because the
user must specifically click the "open" arrow to view the list,
whereas for a non-editable combo they can click anywhere for the
listbox to open.)

Implementation is similar to module paths in the "Go to module" action.
This should make selecting a particular run based on iteration
variable values much more convenient.