qtenv: Make mouse picking of bidirectional connections easier
Both gates will be picked at once at the middle two quarters of
the connection line, instead of just in a square of a small
pixels at the center.
This makes the tooltip and the context menu more useful.
Clicking in the overlapping "middle half" will make the object
inspector target one of the gates arbitrarily.
The appearance (painting) of the connection will still be split
at the middle.